services/applications/vdag/vdag_controller/controller/requirements.txt
grpcio==1.62.1
protobuf==3.20.3
requests==2.32.3
redis==3.2.1
flask==3.0.3
prometheus-client==0.14.1